refactor(api): extract shared pub/sub backend into pubsub.service.ts
